Added and modified some files.

+++ src/Student/Controllers/addMyActivity.controller.js
@@ -10,16 +10,23 @@ const { asyncHandler } = require("../../Admin/Utils/asyncHandler.utils.js");
 const { uploadOnCloudinary } = require("../../Admin/Utils/cloudinary.utils.js");
 
 const addMyActivity = asyncHandler(async (req, res) => {
+  // needed for all (consider this as company name for internship)
+  // needed for all (consider this as domain for internship)
+  // needed for workshop and internship
+  // needed for curricular activities and workshop
+  // needed for project
+  // Number of months internship done
+  // Number of months internship done
   const {
-    title, // needed for all (consider this as company name for internship)
-    message, // needed for all (consider this as domain for internship)
-    place, // needed for workshop and internship
-    date, // needed for curricular activities and workshop
-    facultyId, // needed for project
-    timeperiod, // Number of months internship done
-    isActive, // Number of months internship done
-  } = req.body;
-
+    title,
+    message,
+    place,
+    date,
+    facultyId,
+    guideName,
+    timeperiod,
+    isActive,
+  } = req.body.data || req.body;
   const { type } = req.params;
 
   // Checking if student or faculty
@@ -33,7 +40,7 @@ const addMyActivity = asyncHandler(async (req, res) => {
     if (!fileUrl) {
       throw new ApiError(500, "Failed to upload file to Cloudinary");
     }
-    data.certificateUrl = fileUrl;
+    data.certificateUrl = fileUrl.url;
   }
 
   // Doing the main work of creating document
@@ -55,9 +62,9 @@ const addMyActivity = asyncHandler(async (req, res) => {
 
       if (type === "workshop") {
         let userId;
-        if (req.user.rollNo) {
+        if (req.user?.rollNo) {
           userId = req.user.rollNo;
-        } else if (req.user.facultyId) {
+        } else if (req.user?.facultyId) {
           userId = req.user.facultyId;
         }
 
@@ -77,11 +84,12 @@ const addMyActivity = asyncHandler(async (req, res) => {
         place: record?.place,
         date: record?.date,
       };
-    } else if (type === "project" && req.user.rollNo && facultyId) {
+    } else if (type === "project" && req.user?.rollNo) {
       data.rollNo = req.user.rollNo;
       data = {
         ...data,
         guide: facultyId,
+        guideName,
       };
       record = await project.create(data);
 
@@ -96,12 +104,14 @@ const addMyActivity = asyncHandler(async (req, res) => {
       timeperiod
     ) {
       data = {
+        ...data,
         rollNo: req.user.rollNo,
         companyName: title,
         Domain: message,
         isActive,
         timeperiod,
       };
+
       record = await internship.create(data);
       result = {
         title: record?.companyName,
@@ -142,37 +152,31 @@ const viewActivity = asyncHandler(async (req, res) => {
   }
 
   let result;
-  if (type === "workshop") result = await workShop.findOne(query);
-  else if (type === "internship") result = await internship.findOne(query);
-  else if (type === "curricular-activities")
-    result = await curricularActivities.findOne(query);
-  else if (type === "project") result = await project.findOne(query);
+  if (type === "workshop") result = await workShop.find(query);
+  else if (type === "internship") result = await internship.find(query);
+  else if (type === "curricular-activities") {
+    result = await curricularActivities.find(query);
+  } else if (type === "project") result = await project.find(query);
 
   if (!result)
-    throw new ApiError(500, "Internal Server Error: No details found");
+    return res.status(200).json(new ApiResponse(200, {}, "No details found"));
 
-  res.status(200).json(new ApiResponse(200, result, "Successful"));
+  return res.status(200).json(new ApiResponse(200, result, "Successful"));
 });
 
 const deleteActivity = asyncHandler(async (req, res) => {
   const { ids } = req.body;
 
+  if (!Array.isArray(ids) && ids?.length === 0) {
+    throw new ApiError(404, "Client Error: No details of Activities");
+  }
+
   const { type } = req.params;
 
-  // if (type === "workshop" && req.user?.rollNo) {
-  //   query.userId = req.user.rollNo;
-  // } else if (type === "workshop" && req.user?.facultyId) {
-  //   query.userId = req.user.facultyId;
-  // } else if (type) {
-  //   query.rollNo = req.user.rollNo;
-  // } else {
-  //   throw new ApiError(404, "Client Error: Wrong User");
-  // }
-
   let result;
-  if (type === "workshop")
+  if (type === "workshop") {
     result = await workShop.deleteMany({ _id: { $in: ids } });
-  else if (type === "internship")
+  } else if (type === "internship")
     result = await internship.deleteMany({ _id: { $in: ids } });
   else if (type === "curricular-activities")
     result = await curricularActivities.deleteMany({ _id: { $in: ids } });
